requirements.txt
altgraph==0.16.1
et-xmlfile==1.0.1
future==0.17.1
jdcal==1.4.1
macholib==1.11
numpy==1.16.2
openpyxl==2.6.2
pandas==0.24.2
pefile==2018.8.8
pkg-resources==0.0.0
psycopg2-binary==2.8.1
PyInstaller==3.4
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0

